stage-aris/backend/requirements.txt
Django
pymongo
djangorestframework
